Reduce complexity of SettingsFragment by not overloading responsibility #1945 and make ValidateSearchEngineAsyncTask in seperate class
